pubsub-to-bq/requirements.txt
google-cloud-bigquery
Flask>=2.0
cloudevents>=1.0
